Add REQUEST_ID to build env and return it in Request-Id header
This is setting a REQUEST_ID env for parity with Heroku git pushes so
buildpacks can use this id as a unique value to identify builds.
Currently, it is set to the same as the slug/manifest id for
consistancy, but this could change in the future and shouldn't be
assumed to always be the same.